+++ lib/src/bearing.dart
@@ -9,12 +9,12 @@ num bearingRaw(Position start, Position end, {bool calcFinal = false}) {
     return calculateFinalBearingRaw(start, end);
   }
 
-  num lon1 = degreesToRadians(start.lng);
-  num lon2 = degreesToRadians(end.lng);
+  num lng1 = degreesToRadians(start.lng);
+  num lng2 = degreesToRadians(end.lng);
   num lat1 = degreesToRadians(start.lat);
-  num lat2 = degreesToRadians(start.lat);
-  num a = sin(lon2 - lon1) * cos(lat2);
-  num b = cos(lat1) * sin(lat2) - sin(lat1) * cos(lat2) * cos(lon2 - lon1);
+  num lat2 = degreesToRadians(end.lat);
+  num a = sin(lng2 - lng1) * cos(lat2);
+  num b = cos(lat1) * sin(lat2) - sin(lat1) * cos(lat2) * cos(lng2 - lng1);
 
   return radiansToDegrees(atan2(a, b));
 }