+++ lib/src/destination.dart
@@ -22,7 +22,8 @@ Position destinationRaw(Position origin, num distance, num bearing,
   );
 }
 
-Point destination(Point origin, num distance, num bearing, [Unit unit]) =>
+Point destination(Point origin, num distance, num bearing,
+        [Unit unit = Unit.kilometers]) =>
     Point(
       coordinates: destinationRaw(origin.coordinates, distance, bearing, unit),
     );