+++ test/components/geojson_test.dart
@@ -1,4 +1,7 @@
+import 'dart:math';
+
 import 'package:test/test.dart';
+import 'package:turf/distance.dart';
 import 'package:turf/helpers.dart';
 
 main() {
@@ -43,6 +46,41 @@ main() {
       var bbox2 = BBox.of([1, 2, 3, 4, 5, 6]);
       _expectArgs(bbox1);
       _expectArgs(bbox2);
+    });
+    test('Longitude normalization', () {
+      var rand = Random();
+      _rand() => rand.nextDouble() * (360 * 2) - 360;
+      test('Position.toSigned', () {
+        for (var i = 0; i < 10; i++) {
+          var coord = Position.named(lat: _rand(), lng: _rand(), alt: 0);
+          var zeroZero = Position(0, 0);
+          var distToCoord = distanceRaw(zeroZero, coord);
+          var distToNormalizedCoord = distanceRaw(zeroZero, coord.toSigned());
+
+          expect(distToCoord, distToNormalizedCoord);
+        }
+      });
+      test('BBox.toSigned', () {
+        for (var i = 0; i < 10; i++) {
+          var coord = BBox.named(
+            lat1: _rand(),
+            lat2: _rand(),
+            lng1: _rand(),
+            lng2: _rand(),
+          );
+          var zeroZero = Position(0, 0);
+
+          var distToCoord1 = distanceRaw(
+              zeroZero, Position.named(lng: coord.lng1, lat: coord.lat1));
+          var distToNormalizedCoord1 = distanceRaw(zeroZero, coord.toSigned());
+          expect(distToCoord1, distToNormalizedCoord1);
+
+          var distToCoord2 = distanceRaw(
+              zeroZero, Position.named(lng: coord.lng2, lat: coord.lat2));
+          var distToNormalizedCoord2 = distanceRaw(zeroZero, coord.toSigned());
+          expect(distToCoord2, distToNormalizedCoord2);
+        }
+      });
     });
     test('Point', () {});
     test('Point', () {});