+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- normalization for coordinates (Position)
+- and yes, it's still under heavy development
+
 ## 0.0.1
 
 - Initial version, still under heavy development